chore: Update MigrateProductCard layout for responsive design
