+++ apps/rays-dashboard/components/molecules/MigrateProductCard/MigrateProductCard.module.scss
@@ -3,7 +3,7 @@
   display: flex;
   flex-direction: column;
 
-  @media screen and (min-width: 600px) {
+  @include media('<=s') {
     min-width: 818px;
   }
 }
