+++ apps/rays-dashboard/components/molecules/MigrateProductCard/MigrateProductCardProtocolInfo.tsx
@@ -1,0 +1,26 @@
+import { type TokenSymbolsList } from '@summerfi/app-types'
+import { Text, TokensGroup } from '@summerfi/app-ui'
+
+import { type LendingProtocolConfig } from '@/helpers/lending-protocols-configs'
+
+import migrateProductCardStyles from '@/components/molecules/MigrateProductCard/MigrateProductCard.module.scss'
+
+export const MigrateProductCardProtocolInfo = ({
+  tokens,
+  protocolConfig,
+}: {
+  tokens: TokenSymbolsList[]
+  protocolConfig: LendingProtocolConfig
+}) => (
+  <div className={migrateProductCardStyles.protocolInfoWrapper}>
+    <TokensGroup tokens={tokens} />
+    <div className={migrateProductCardStyles.protocolInfo}>
+      <Text as="p" variant="p1semi">
+        {tokens.join('/')}
+      </Text>
+      <Text as="p" variant="p3" style={{ color: 'var(--color-neutral-80)' }}>
+        {protocolConfig.label}
+      </Text>
+    </div>
+  </div>
+)
