+++ apps/rays-dashboard/components/molecules/RaysProductCard/RaysProductCard.module.scss
@@ -1,7 +1,7 @@
 .cardWrapper {
   width: auto;
 
-  @media screen and (min-width: 600px) {
+  @include media('<=s') {
     min-width: 818px;
   }
 }
