+++ apps/rays-dashboard/components/organisms/Leaderboard/LeaderboardSearchBoxAndResults.module.scss
@@ -5,7 +5,7 @@
     justify-content: flex-end;
     width: 100%;
     position: relative;
-    @media (min-width: 744px) {
+    @include media('<=m') {
       margin: initial;
       margin-bottom: -65px;
     }
